Extract an interface for quality and name the implementation so that it reveals some of the implementation details
