+++ src/main/java/org/agilegrenoble/objectcalisthenics/ItemForge.java
@@ -6,13 +6,13 @@ import org.agilegrenoble.objectcalisthenics.Ageing.ImproveWithAge;
 import org.agilegrenoble.objectcalisthenics.items.Item;
 import org.agilegrenoble.objectcalisthenics.items.NormalItem;
 import org.agilegrenoble.objectcalisthenics.items.Sulfuras;
-import org.agilegrenoble.objectcalisthenics.quality.Quality;
+import org.agilegrenoble.objectcalisthenics.quality.Between0and50Quality;
 
 public class ItemForge {
 
     public Item anAgedBrie_thatImprovesWithTime(int daysUntilBestBeforeDate, int startQuality) {
         return new NormalItem("Aged Brie", 
-                new ImproveWithAge(daysUntilBestBeforeDate), new Quality(startQuality));
+                new ImproveWithAge(daysUntilBestBeforeDate), new Between0and50Quality(startQuality));
     }
 
     public Item aSulfuras_isAMagicItemThatNeverChanges() {
@@ -21,12 +21,12 @@ public class ItemForge {
 
     public Item anItem_thatDecaysWithTime(String name, int daysUntilBestBeforeDate, int startQuality) {
         return new NormalItem(name, 
-                new DecayWithAge(daysUntilBestBeforeDate), new Quality(startQuality));
+                new DecayWithAge(daysUntilBestBeforeDate), new Between0and50Quality(startQuality));
     }
 
     public Item aBackstagePass_thatImprovesUntilTheConcertDate(int daysUntilConcert, int startQuality) {
         return new NormalItem("Backstage passes to a TAFKAL80ETC concert", 
-                new BackStagePassAgeing(daysUntilConcert), new Quality(startQuality));
+                new BackStagePassAgeing(daysUntilConcert), new Between0and50Quality(startQuality));
     }
 
 }
