+++ src/main/java/org/agilegrenoble/objectcalisthenics/Main.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.agilegrenoble.objectcalisthenics.Ageing.DecayWithAge;
 import org.agilegrenoble.objectcalisthenics.items.Item;
 import org.agilegrenoble.objectcalisthenics.items.NormalItem;
-import org.agilegrenoble.objectcalisthenics.quality.Quality;
+import org.agilegrenoble.objectcalisthenics.quality.Between0and50Quality;
 
 public class Main {
 
@@ -19,8 +19,11 @@ public class Main {
 		items.add(forge.aBackstagePass_thatImprovesUntilTheConcertDate(15, 20));
 		items.add(forge.anItem_thatDecaysWithTime("+5 Dexterity Vest", 10, 20));
 		items.add(forge.anItem_thatDecaysWithTime("Elixir of the Mongoose", 5, 7));
-		items.add(new NormalItem("Conjured Mana Cake", new DecayWithAge(3), new Quality(6)));
+		items.add(new NormalItem("Conjured Mana Cake", new DecayWithAge(3), new Between0and50Quality(6)));
 		new Inn(items).updateQuality();
 	}
+	
+	// daysUntilPerumption/perdition? for normal items
+	// daysUntilFirstConsumptionDate?
 
 }
